+++ gems/aws-sdk-s3/lib/aws-sdk-s3/multipart_upload_error.rb
@@ -2,17 +2,16 @@
 
 module Aws
   module S3
+    # Raise when multipart upload fails to complete.
     class MultipartUploadError < StandardError
 
-      def initialize(message, errors)
+      def initialize(message, errors = [])
         @errors = errors
         super(message)
       end
 
-      # @return [Array<StandardError>] The list of errors encountered
-      #   when uploading or aborting the upload.
+      # @return [Array<StandardError>] The list of errors encountered when uploading or aborting the upload.
       attr_reader :errors
-
     end
   end
 end